First pass at converting chemical amounts to be longs instead of ints in preparation for eventual changes to mass amounts of steam production - Various minor config fixes/adjustments, highly recommended to regenerate configs to ensure things work properly - Improve networking for various things by using varInt rather than int, as the locations we pass ints over networks they tend to be small - Partially rewrite ProjectE integration to better handle/fail gracefully if a recipe has a long amount of a chemical, and properly handle if multiple inputs are the same

+++ src/api/java/mekanism/api/chemical/gas/GasHandlerWrapper.java
@@ -36,7 +36,7 @@ public class GasHandlerWrapper implements IChemicalHandlerWrapper<Gas, GasStack>
     }
 
     @Override
-    public int getTankCapacity(int tank) {
+    public long getTankCapacity(int tank) {
         return handler.getGasTankCapacity(tank);
     }
 
@@ -51,7 +51,7 @@ public class GasHandlerWrapper implements IChemicalHandlerWrapper<Gas, GasStack>
     }
 
     @Override
-    public GasStack extractChemical(int tank, int amount, Action action) {
+    public GasStack extractChemical(int tank, long amount, Action action) {
         return handler.extractGas(tank, amount, action);
     }
 
@@ -61,7 +61,7 @@ public class GasHandlerWrapper implements IChemicalHandlerWrapper<Gas, GasStack>
     }
 
     @Override
-    public GasStack extractChemical(int amount, Action action) {
+    public GasStack extractChemical(long amount, Action action) {
         return handler.extractGas(amount, action);
     }
