+++ src/main/java/mekanism/client/jei/chemical/GasStackRenderer.java
@@ -12,7 +12,7 @@ public class GasStackRenderer extends ChemicalStackRenderer<Gas, GasStack> {
         super(FluidAttributes.BUCKET_VOLUME, TooltipMode.ITEM_LIST, TEX_WIDTH, TEX_HEIGHT, null);
     }
 
-    public GasStackRenderer(int capacityMb, boolean showCapacity, int width, int height, @Nullable IDrawable overlay) {
+    public GasStackRenderer(long capacityMb, boolean showCapacity, int width, int height, @Nullable IDrawable overlay) {
         super(capacityMb, showCapacity ? TooltipMode.SHOW_AMOUNT_AND_CAPACITY : TooltipMode.SHOW_AMOUNT, width, height, overlay);
     }
 }
