+++ src/main/java/mekanism/common/capabilities/chemical/MultiblockGasTank.java
@@ -2,7 +2,7 @@ package mekanism.common.capabilities.chemical;
 
 import java.util.Objects;
 import java.util.function.BiPredicate;
-import java.util.function.IntSupplier;
+import java.util.function.LongSupplier;
 import java.util.function.Predicate;
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
@@ -20,18 +20,18 @@ import mekanism.common.tile.TileEntityMultiblock;
 @MethodsReturnNonnullByDefault
 public class MultiblockGasTank<MULTIBLOCK extends TileEntityMultiblock<?>> extends VariableCapacityGasTank {
 
-    public static <MULTIBLOCK extends TileEntityMultiblock<?>> MultiblockGasTank<MULTIBLOCK> create(MULTIBLOCK multiblock, IntSupplier capacity, Predicate<@NonNull Gas> validator) {
+    public static <MULTIBLOCK extends TileEntityMultiblock<?>> MultiblockGasTank<MULTIBLOCK> create(MULTIBLOCK multiblock, LongSupplier capacity, Predicate<@NonNull Gas> validator) {
         return new MultiblockGasTank<>(multiblock, capacity, validator);
     }
 
-    public static <MULTIBLOCK extends TileEntityMultiblock<?>> MultiblockGasTank<MULTIBLOCK> create(MULTIBLOCK multiblock, IntSupplier capacity,
+    public static <MULTIBLOCK extends TileEntityMultiblock<?>> MultiblockGasTank<MULTIBLOCK> create(MULTIBLOCK multiblock, LongSupplier capacity,
           BiPredicate<@NonNull Gas, @NonNull AutomationType> canExtract, BiPredicate<@NonNull Gas, @NonNull AutomationType> canInsert, Predicate<@NonNull Gas> validator) {
         return new MultiblockGasTank<>(multiblock, capacity, canExtract, canInsert, validator, null, null);
     }
 
-    public static <MULTIBLOCK extends TileEntityMultiblock<?>> MultiblockGasTank<MULTIBLOCK> create(MULTIBLOCK tile, IntSupplier capacity, BiPredicate<@NonNull Gas, @NonNull AutomationType> canExtract,
-          BiPredicate<@NonNull Gas, @NonNull AutomationType> canInsert, Predicate<@NonNull Gas> validator, @Nullable ChemicalAttributeValidator attributeValidator,
-          @Nullable IMekanismGasHandler gasHandler) {
+    public static <MULTIBLOCK extends TileEntityMultiblock<?>> MultiblockGasTank<MULTIBLOCK> create(MULTIBLOCK tile, LongSupplier capacity,
+          BiPredicate<@NonNull Gas, @NonNull AutomationType> canExtract, BiPredicate<@NonNull Gas, @NonNull AutomationType> canInsert, Predicate<@NonNull Gas> validator,
+          @Nullable ChemicalAttributeValidator attributeValidator, @Nullable IMekanismGasHandler gasHandler) {
         Objects.requireNonNull(tile, "Tile cannot be null");
         Objects.requireNonNull(capacity, "Capacity supplier cannot be null");
         Objects.requireNonNull(validator, "Gas validity check cannot be null");
@@ -42,13 +42,13 @@ public class MultiblockGasTank<MULTIBLOCK extends TileEntityMultiblock<?>> exten
 
     protected final MULTIBLOCK multiblock;
 
-    protected MultiblockGasTank(MULTIBLOCK multiblock, IntSupplier capacity, Predicate<@NonNull Gas> validator) {
+    protected MultiblockGasTank(MULTIBLOCK multiblock, LongSupplier capacity, Predicate<@NonNull Gas> validator) {
         super(capacity, (stack, automationType) -> automationType != AutomationType.EXTERNAL || multiblock.structure != null,
             (stack, automationType) -> automationType != AutomationType.EXTERNAL || multiblock.structure != null, validator, null, null);
         this.multiblock = multiblock;
     }
 
-    protected MultiblockGasTank(MULTIBLOCK multiblock, IntSupplier capacity, BiPredicate<@NonNull Gas, @NonNull AutomationType> canExtract,
+    protected MultiblockGasTank(MULTIBLOCK multiblock, LongSupplier capacity, BiPredicate<@NonNull Gas, @NonNull AutomationType> canExtract,
           BiPredicate<@NonNull Gas, @NonNull AutomationType> canInsert, Predicate<@NonNull Gas> validator, @Nullable ChemicalAttributeValidator attributeValidator,
           @Nullable IMekanismGasHandler gasHandler) {
         super(capacity, canExtract, canInsert, validator, attributeValidator, gasHandler);
